Release 3.1.0

+++ src/main/java/io/penguinstats/model/query/TrendQuery.java
@@ -43,7 +43,7 @@ public class TrendQuery implements Serializable, BasicQuery {
 
 	private String userID;
 
-	private Integer interval;
+	private Long interval;
 
 	private Integer timeout;
